requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
boto3==1.38.21
botocore==1.38.21
certifi==2025.4.26
charset-normalizer==3.4.2
diffusers==0.33.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.0
h11==0.16.0
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
lightning-utilities==0.14.3
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.12.0
safetensors==0.5.3
selenium==4.32.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.52.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
